Add cui::panels::tab_stack namespace
